+++ coc/extensions/coc-sumneko-lua-data/sumneko-lua-ls/extension/server/script/core/diagnostics/unused-function.lua
@@ -18,6 +18,7 @@ local function isToBeClosed(source)
     return false
 end
 
+---@async
 return function (uri, callback)
     local ast = files.getState(uri)
     if not ast then
@@ -25,6 +26,7 @@ return function (uri, callback)
     end
 
     local cache = {}
+    ---@async
     local function checkFunction(source)
         if cache[source] ~= nil then
             return cache[source]
@@ -81,7 +83,7 @@ return function (uri, callback)
     end
 
     -- 只检查局部函数
-    guide.eachSourceType(ast.ast, 'function', function (source)
+    guide.eachSourceType(ast.ast, 'function', function (source) ---@async
         checkFunction(source)
     end)
 end