+++ coc/extensions/coc-sumneko-lua-data/sumneko-lua-ls/extension/server/script/workspace/workspace.lua
@@ -74,6 +74,7 @@ local globInteferFace = {
 }
 
 --- 创建排除文件匹配器
+---@async
 function m.getNativeMatcher()
     if not m.path then
         return nil
@@ -130,7 +131,7 @@ function m.getNativeMatcher()
         end
     end
     -- config.get 'workspace.ignoreDir'
-    for path in pairs(config.get 'Lua.workspace.ignoreDir') do
+    for _, path in ipairs(config.get 'Lua.workspace.ignoreDir') do
         log.info('Ignore directory:', path)
         pattern[#pattern+1] = path
     end
@@ -177,6 +178,7 @@ function m.getLibraryMatchers()
 end
 
 --- 文件是否被忽略
+---@async
 function m.isIgnored(uri)
     local path = m.getRelativePath(uri)
     local ignore = m.getNativeMatcher()
@@ -186,6 +188,7 @@ function m.isIgnored(uri)
     return ignore(path)
 end
 
+---@async
 function m.isValidLuaUri(uri)
     if not files.isLua(uri) then
         return false
@@ -198,7 +201,7 @@ function m.isValidLuaUri(uri)
 end
 
 local function loadFileFactory(root, progressData, isLibrary)
-    return function (path)
+    return function (path) ---@async
         local uri = furi.encode(path)
         if files.isLua(uri) then
             if not isLibrary and progressData.preload >= config.get 'Lua.workspace.maxPreload' then
@@ -246,7 +249,7 @@ local function loadFileFactory(root, progressData, isLibrary)
                             log.info('++++As library of:', root)
                             files.setLibraryPath(uri, root)
                         end
-                        files.setText(uri, text, false, true)
+                        files.setText(uri, text, false)
                     else
                         files.remove(uri)
                     end
@@ -279,6 +282,7 @@ local function loadFileFactory(root, progressData, isLibrary)
     end
 end
 
+---@async
 function m.awaitLoadFile(uri)
     local progressBar <close> = progress.create(lang.script.WORKSPACE_LOADING)
     local progressData = {
@@ -299,6 +303,7 @@ function m.awaitLoadFile(uri)
 end
 
 --- 预读工作区内所有文件
+---@async
 function m.awaitPreload()
     local diagnostic = require 'provider.diagnostic'
     await.close 'preload'
@@ -347,7 +352,7 @@ function m.awaitPreload()
         if isLoadingFiles then
             return
         end
-        await.call(function ()
+        await.call(function () ---@async
             isLoadingFiles = true
             while true do
                 local loader = table.remove(progressData.loaders)
@@ -391,17 +396,22 @@ function m.findUrisByFilePath(path)
         return resultCache[path].results, resultCache[path].posts
     end
     tracy.ZoneBeginN('findUrisByFilePath #1')
+    local strict = config.get 'Lua.runtime.pathStrict'
     local results = {}
     local posts = {}
     for uri in files.eachFile() do
         if not uri:find(lpath, 1, true) then
             goto CONTINUE
         end
+        local relat = m.getRelativePath(uri)
         local pathLen = #path
-        local curPath = furi.decode(uri)
+        local curPath = relat
         local curLen  = #curPath
         local seg = curPath:sub(curLen - pathLen, curLen - pathLen)
         if seg == '/' or seg == '\\' or seg == '' then
+            if strict and seg ~= '' then
+                goto CONTINUE
+            end
             local see = curPath:sub(curLen - pathLen + 1, curLen)
             if see == path then
                 results[#results+1] = uri
@@ -471,6 +481,10 @@ function m.normalize(path)
     path = path:gsub('%$%{(.-)%}', function (key)
         if key == '3rd' then
             return (ROOT / 'meta' / '3rd'):string()
+        end
+        if key:sub(1, 4) == 'env:' then
+            local env = os.getenv(key:sub(5))
+            return env
         end
     end)
     path = util.expandPath(path)
@@ -558,6 +572,7 @@ function m.init()
     m.reload()
 end
 
+---@async
 function m.awaitReload()
     m.ready = false
     m.hasHitMaxPreload = false
@@ -575,6 +590,7 @@ function m.awaitReload()
 end
 
 ---等待工作目录加载完成
+---@async
 function m.awaitReady()
     if m.isReady() then
         return
@@ -592,7 +608,7 @@ function m.getLoadProcess()
     return m.fileLoaded, m.fileFound
 end
 
-files.watch(function (ev, uri)
+files.watch(function (ev, uri) ---@async
     if  ev == 'close'
     and m.isIgnored(uri)
     and not files.isLibrary(uri) then
@@ -610,7 +626,7 @@ config.watch(function (key, value, oldValue)
     end
 end)
 
-fw.event(function (changes)
+fw.event(function (changes) ---@async
     m.awaitReady()
     for _, change in ipairs(changes) do
         local path = change.path