+++ nvim__/init.lua
@@ -1,4 +1,0 @@
-require('plugins')
-require('keymapping')
-require('nv-compe')
-require('settings')